Recur on nested sidebar layers
